refactor: reorganize network utilities and parsing logic, separate sorting and parsing into distinct modules
